pyproject.toml
[project]
name = "kok_products_ble"
version = "0.1.0"
description = "Integration for Home Assistant OS which allows control of my personal BLE devices"
requires-python = ">=3.13"
dependencies = [
    "pydantic==2.12.5",
]

[dependency-groups]
dev = [
    "homeassistant-stubs>=2025.4.4",
]
